requirements.txt
streamlit>=1.31
openai>=1.12
python-dotenv>=1.0
